+++ src/controllers/baseinfo/depart.js
@@ -10,16 +10,21 @@ export const list = asyncHandler(async (req, res, next) => {
   const pagesize = parseInt(req.query.pagesize ?? 10)
   const query = req.query.query ? JSON.parse(req.query.query) : ''
   
+  const myQuery = {}
   // 处理查询条件
   if (isDef(query.departName)) {
-    query.departName = {
+    myQuery.departName = {
       [Op.startsWith]: query.departName
     }
   }
 
+  if (query.campusId && query.campusId !== -1) {
+    myQuery.CampusId = query.campusId
+  }
+
   const data = await Depart.findAndCountAll({
     order: [[ 'id', 'DESC' ]],
-    where: query,
+    where: myQuery,
     include: [
       { model: Campus }
     ],
@@ -50,6 +55,7 @@ export const list = asyncHandler(async (req, res, next) => {
 })
 
 export const add = asyncHandler(async (req, res, next) => {
+  req.body.CampusId = req.body.campusId
   const data = await Depart.create(req.body)
   res.status(201).json({
     code: 201,
@@ -60,6 +66,7 @@ export const add = asyncHandler(async (req, res, next) => {
 
 export const update = asyncHandler(async (req, res, next) => {
   const id = req.params.id
+  req.body.CampusId = req.body.campusId
   const data = await Depart.update(req.body, {
     where: {
       id: id
@@ -74,6 +81,19 @@ export const update = asyncHandler(async (req, res, next) => {
 
 export const del = asyncHandler(async (req, res, next) => {
   const id = req.params.id
+
+  const res1 = await Employee.findAndCountAll({
+    where: {
+      CampusId: id
+    }
+  })
+  if (res1.count > 0) {
+    return res.status(401).json({
+      code: 401,
+      message: '该部门下有员工，请先删除员工'
+    })
+  }
+
   const data = await Depart.destroy({
     where: {
       id
